addIcon moved to helper.js
