+++ src/modules/factoryFunc.js
@@ -1,5 +1,4 @@
 import { modal, openModal } from './helper';
-// import * as UI from './ui';
 
 const newListInput = document.getElementById('sidebar-input');
