--- src/modules/ui.js
@@ -1,6 +1,0 @@
-
-export const modal = document.querySelector('.modal');
-export const advButton = document.querySelector('.advButton');
-export const advOptions = document.querySelector('.advanced-options');
-
-export const newListInput = document.getElementById('sidebar-input');
